refactor: move methods around into better named files
